refactored structure of network classes
